[cli-tests] Add help test for wait-ready
Also move locale availability code to its own entity.
