+++ tools/cli_tests/multipass/feature_versions.py
@@ -28,6 +28,7 @@ def _feature_version(name):
         "blueprints": {"max": ver.parse("1.16")},
         "debian_images": {"min": ver.parse("1.17")},
         "fedora_images": {"min": ver.parse("1.17")},
+        "wait_ready": {"min": ver.parse("1.17")},
     }
     assert name in feats, f"No such feature: {name}"
     return feats[name]
@@ -51,6 +52,7 @@ def multipass_version_has_feature(feature_name, version=None):
             return False
 
     return True
+
 
 def test_requires_feature(feature_name, version=None):
     if not multipass_version_has_feature(feature_name, version=version):
