+++ tools/cli_tests/utilities/localeutils.py
@@ -1,0 +1,40 @@
+#!/usr/bin/env python3
+#
+# Copyright (C) Canonical, Ltd.
+#
+# This program is free software; you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation; version 3.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+#
+
+from .privutils import run_in_new_interpreter
+
+
+def set_locale(loc):
+    import locale
+
+    locale.setlocale(locale.LC_ALL, loc)
+
+
+def is_locale_available(loc):
+    """Intended to be run in a subprocess."""
+    import subprocess
+
+    return (
+        run_in_new_interpreter(
+            set_locale,
+            loc,
+            check=False,
+            stdout=subprocess.DEVNULL,
+            stderr=subprocess.DEVNULL,
+        ).returncode
+        == 0
+    )
